+++ packages/sdk/src/sdk-blockchains/ethereum/sell.ts
@@ -4,6 +4,7 @@ import type { EthereumNetwork } from "@rarible/protocol-ethereum-sdk/build/types
 import type * as OrderCommon from "../../types/order/common"
 import { OriginFeeSupport, PayoutsSupport } from "../../types/order/fill/domain"
 import { getCurrencyAssetType } from "../../common/get-currency-asset-type"
+import type { PrepareSellInternalResponse } from "../../types/order/sell/domain"
 import * as common from "./common"
 import type { EVMBlockchain } from "./common"
 import { getEthereumItemId, getEVMBlockchain, isEVMBlockchain } from "./common"
@@ -20,15 +21,7 @@ export class EthereumSell {
 		this.update = this.update.bind(this)
 	}
 
-	async sell(request: OrderCommon.PrepareOrderInternalRequest): Promise<OrderCommon.PrepareOrderInternalResponse> {
-		const [domain, contract] = request.collectionId.split(":")
-		if (!isEVMBlockchain(domain)) {
-			throw new Error("Not an ethereum item")
-		}
-		const collection = await this.sdk.apis.nftCollection.getNftCollectionById({
-			collection: contract,
-		})
-
+	async sell(): Promise<PrepareSellInternalResponse> {
 		const sellAction = this.sdk.order.sell
 			.before(async (sellFormRequest: OrderCommon.OrderInternalRequest) => {
 				const { itemId } = getEthereumItemId(sellFormRequest.itemId)
@@ -55,7 +48,6 @@ export class EthereumSell {
 		return {
 			originFeeSupport: OriginFeeSupport.FULL,
 			payoutsSupport: PayoutsSupport.MULTIPLE,
-			multiple: collection.type === "ERC1155",
 			supportedCurrencies: common.getSupportedCurrencies(),
 			baseFee: await this.sdk.order.getBaseOrderFee(),
 			supportsExpirationDate: true,