+++ packages/sdk/src/sdk-blockchains/ethereum/test/await-balance.ts
@@ -1,14 +1,14 @@
 import BigNumber from "bignumber.js"
-import type { AssetType } from "@rarible/api-client"
 import { Blockchain } from "@rarible/api-client"
 import type { EthereumWallet } from "@rarible/sdk-wallet"
 import type { BigNumberValue } from "@rarible/utils"
 import { convertEthereumToUnionAddress } from "../common"
 import type { IRaribleSdk } from "../../../domain"
 import { retry } from "../../../common/retry"
+import type { RequestCurrencyAssetType } from "../../../common/domain"
 
 export async function awaitBalance(
-	sdk: IRaribleSdk, assetType: AssetType, wallet: EthereumWallet, value: BigNumberValue
+	sdk: IRaribleSdk, assetType: RequestCurrencyAssetType, wallet: EthereumWallet, value: BigNumberValue
 ) {
 	return await retry(5, 2000, async () => {
 		const address = convertEthereumToUnionAddress(await wallet.ethereum.getFrom(), Blockchain.ETHEREUM)