+++ packages/sdk/src/sdk-blockchains/flow/balance.ts
@@ -1,11 +1,12 @@
 import type { UnionAddress } from "@rarible/types"
-import type { AssetType } from "@rarible/api-client"
 import type { BigNumberValue } from "@rarible/utils"
 import type { FlowSdk } from "@rarible/flow-sdk"
 import { toBn } from "@rarible/utils/build/bn"
 import type { Maybe } from "@rarible/types/build/maybe"
 import type { FlowWallet } from "@rarible/sdk-wallet"
 import type { FlowEnv } from "@rarible/flow-sdk/build/types"
+import { getCurrencyAssetType } from "../../common/get-currency-asset-type"
+import type { RequestCurrency } from "../../common/domain"
 import { parseFlowAddressFromUnionAddress } from "./common/converters"
 import { getFlowCurrencyFromAssetType } from "./common/get-flow-currency-from-asset-type"
 import { getSimpleFlowFungibleBalance } from "./balance-simple"
@@ -19,7 +20,8 @@ export class FlowBalance {
 		this.getBalance = this.getBalance.bind(this)
 	}
 
-	async getBalance(address: UnionAddress, assetType: AssetType): Promise<BigNumberValue> {
+	async getBalance(address: UnionAddress, currency: RequestCurrency): Promise<BigNumberValue> {
+		const assetType = getCurrencyAssetType(currency)
 		if (this.wallet) {
 			const flowAddress = parseFlowAddressFromUnionAddress(address)
 			const flowAsset = getFlowCurrencyFromAssetType(assetType)